requirements.txt
httpx-oauth==0.13.2
google-api-python-client==2.121.0
google-auth==2.28.2
google-auth-oauthlib==1.2.0
st-annotated-text==4.0.1
streamlit-modal==0.1.2